+++ src/queries/index.js
@@ -4,4 +4,5 @@ export * from './useAuthorityLinksCount';
 export * from './useLinkSuggestions';
 export * from './useMarcRecordMutation';
 export * from './useAuthorityFileNextHrid';
+export * from './useLccnDuplicateConfig';
 export * from './useValidate';